Ido and Shays CR
